requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
APScheduler==3.10.4
beautifulsoup4==4.12.3
kafka-python==2.0.2
numpy==2.1.1
pandas==2.2.2
pre-commit==3.8.0
requests==2.32.3
SQLAlchemy==2.0.36
pytz~=2024.1